requirements.txt
altair==4.2.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
Bottleneck==1.3.5
branca==0.6.0
brotlipy==0.7.0
cachetools==4.2.2
category-encoders==2.5.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
folium==0.13.0
fonttools==4.38.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
idna==3.4
importlib-metadata==4.11.3
ipykernel==6.15.2
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.1
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter_core==4.11.2
jupyter-server==1.18.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-service==2.4.0
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.5.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.3
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
regex==2022.7.9
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
Send2Trash==1.8.0
setuptools==65.5.0
six==1.16.0
sklearn==0.0.post1
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.15.0
streamlit-folium==0.7.0
terminado==0.13.1
threadpoolctl==2.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.3.0
tzlocal==2.1
urllib3==1.26.12
validators==0.18.2
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==3.5.2
wordcloud==1.8.2.2
zipp==3.8.0
